requirements.txt
markitdown[pptx,docx,pdf]>=0.1.1
openai>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
tiktoken>=0.3.0
tabulate>=0.9.0
pandas>=2.0.0
openpyxl>=3.1.5